fix: add ownership check for user account

+++ src/modules/users/users.service.ts
@@ -1,21 +1,19 @@
 /* eslint-disable prettier/prettier */
-import {
-  ConflictException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { randomUUID } from 'node:crypto';
 import { MailService } from '../utils/mail.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/users.repository';
 
+/**
+ * Service responsible for user manipulation.
+ *
+ * This service handles operations related to users, such as updating and removing.
+ */
 @Injectable()
 export class UsersService {
-  constructor(
-    private userRepository: UsersRepository,
-    private mailService: MailService,
-  ) {}
+  constructor(private userRepository: UsersRepository, private mailService: MailService) {}
 
   async create(createUserDto: CreateUserDto) {
     const findUser = await this.userRepository.findByEmail(createUserDto.email);
@@ -62,14 +60,18 @@ export class UsersService {
     return findUser;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(reqId: string, id: string, updateUserDto: UpdateUserDto) {
+    if (reqId !== id) throw new ForbiddenException('Insufficient permission');
+
     const findUser = await this.userRepository.findOne(id);
     if (!findUser) throw new NotFoundException('User not found');
 
     return this.userRepository.update(id, updateUserDto);
   }
 
-  async remove(id: string) {
+  async remove(reqId: string, id: string) {
+    if (reqId !== id) throw new ForbiddenException('Insufficient permission');
+
     const findUser = await this.userRepository.findOne(id);
     if (!findUser) throw new NotFoundException('User not found');
     return this.userRepository.delete(id);
@@ -84,11 +86,7 @@ export class UsersService {
 
     await this.userRepository.updateToken(email, resetToken);
 
-    const resetPasswordTemplate = await this.mailService.resetPasswordTemplate(
-      email,
-      user.name,
-      resetToken,
-    );
+    const resetPasswordTemplate = await this.mailService.resetPasswordTemplate(email, user.name, resetToken);
     await this.mailService.sendEMail(resetPasswordTemplate);
   }
